+++ apps/mobile/src/assets/icons/index.ts
@@ -12,6 +12,7 @@ export { EyeOffIcon } from "./icons/eye-off.icon";
 export { ChevronLeftIcon } from "./icons/chevron-left.icon";
 export { ChevronDownIcon } from "./icons/chevron-down.icon";
 export { MapPinIcon } from "./icons/map-pin.icon";
+export { NavigationIcon } from "./icons/navigation.icon";
 export { SlidersHorizontalIcon } from "./icons/sliders-horizontal.icon";
 export { HeartIcon } from "./icons/heart.icon";
 export { ClockIcon } from "./icons/clock.icon";